lab3-4/homework/I2/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "zeroc-ice>=3.7.10.1",
]
